rename chat.tsx to index.tsx
